Move CachedOperation to a new file
